Added an extra check in the Docker build scripts to ensure they are called from the right plaace and added the option to disable pushing the latest tag in the push script.

+++ docker/tag_push_multiarch.sh
@@ -29,21 +29,28 @@ docker manifest push docker.io/$USER/timeseria:$VERSION
 
 #============================================================
 
-echo "Tagging latest..."
-docker tag timeseria-amd64 docker.io/$USER/timeseria-amd64:latest
-docker tag timeseria-arm64v8 docker.io/$USER/timeseria-arm64v8:latest
+if [[ "x$TAG_LATEST" == "xTrue" ]]; then
 
-echo "Pushing images for latest"
-docker push docker.io/$USER/timeseria-amd64:latest
-docker push docker.io/$USER/timeseria-arm64v8:latest
+    echo "Tagging latest..."
+    docker tag timeseria-amd64 docker.io/$USER/timeseria-amd64:latest
+    docker tag timeseria-arm64v8 docker.io/$USER/timeseria-arm64v8:latest
 
-echo "Creating manifest for latest..."
-# https://unix.stackexchange.com/questions/393310/preventing-shell-from-exiting-when-set-e-is-turned-on
-docker manifest rm docker.io/$USER/timeseria:latest || true
-docker manifest create \
-docker.io/$USER/timeseria:latest \
---amend docker.io/$USER/timeseria-amd64:latest \
---amend docker.io/$USER/timeseria-arm64v8:latest
+    echo "Pushing images for latest"
+    docker push docker.io/$USER/timeseria-amd64:latest
+    docker push docker.io/$USER/timeseria-arm64v8:latest
 
-echo "Pushing manifest for latest"
-docker manifest push docker.io/$USER/timeseria:latest
+    echo "Creating manifest for latest..."
+    # https://unix.stackexchange.com/questions/393310/preventing-shell-from-exiting-when-set-e-is-turned-on
+    docker manifest rm docker.io/$USER/timeseria:latest || true
+    docker manifest create \
+    docker.io/$USER/timeseria:latest \
+    --amend docker.io/$USER/timeseria-amd64:latest \
+    --amend docker.io/$USER/timeseria-arm64v8:latest
+
+    echo "Pushing manifest for latest"
+    docker manifest push docker.io/$USER/timeseria:latest
+
+else
+    echo "Not tagging latest, use TAG_LATEST=True if you want so"
+
+fi
